// packages/plugins/TransformControlsPlugin/TransformControlsPlane.d.ts
import { Matrix4, Mesh, Object3D, OrthographicCamera, PerspectiveCamera, Quaternion, Vector3 } from "three";
export declare class TransformControlsPlane extends Mesh {
    camera: PerspectiveCamera | OrthographicCamera;
    object?: Object3D;
    enabled: boolean;
    mode: "scale" | "position" | "rotation";
    space: "local" | "world";
    gizmo: any;
    picker: any;
    helper: any;
    axis: string;
    translationSnap: null;
    rotationSnap: null;
    scaleSnap: null;
    size: number;
    dragging: boolean;
    showX: boolean;
    showY: boolean;
    showZ: boolean;
    worldPosition: Vector3;
    worldPositionStart: Vector3;
    worldQuaternion: Quaternion;
    worldQuaternionStart: Quaternion;
    cameraPosition: Vector3;
    cameraQuaternion: Quaternion;
    pointStart: Vector3;
    pointEnd: Vector3;
    rotationAxis: Vector3;
    rotationAngle: number;
    eye: Vector3;
    _tempVector: Vector3;
    _identityQuaternion: Quaternion;
    _alignVector: Vector3;
    _dirVector: Vector3;
    _tempMatrix: Matrix4;
    _unitX: Vector3;
    _unitY: Vector3;
    _unitZ: Vector3;
    _v1: Vector3;
    _v2: Vector3;
    _v3: Vector3;
    constructor();
    updateMatrixWorld(force: any): void;
}